Abstracted the edit and create page
