+++ app/thread/textEditor.tsx
